// packages/vreo/resources/typings/VreoUnit.ts
export enum VreoKeyframeEnum {
  /** 运镜 */
  CameraMovement = 'CameraMovement',
  /**
   * 全景标签
   */
  PanoTag = 'PanoTag',
  /**
   * 全景文本标签
   */
  PanoTextLabel = 'PanoTextLabel',
  /**
   * 提词器
   */
  Prompter = 'Prompter',
  /**
   * 更新 VR 全景
   */
  UpdateVRPanorama = 'UpdateVRPanorama',
  /**
   * 视频广告
   */
  ModelVideo = 'ModelVideo',
  /**
   * 全景特效
   */
  PanoEffect = 'PanoEffect',
  /**
   * 信息面板
   */
  InfoPanel = 'InfoPanel',
  /**
   * 视频特效
   */
  VideoEffect = 'VideoEffect',
  /**
   * 背景音乐
   */
  BgMusic = 'BgMusic',
  /**
   * 结束
   */
  Exit = 'Exit',
  /**
   * 自定义剧本帧
   */
  Custom = 'Custom',
}

/**
 * 支持自定义剧本关键帧配置项
 */
export type VreoKeyframeConfigMap = {
  PanoTag?: false
  PanoTextLabel?: false
  Prompter?: false
  InfoPanel?: false
}


/**
 * 剧本关键帧
 */
export interface VreoKeyframe {
  uuid?: string
  type: VreoKeyframeEnum
  start: number
  end: number
  parsed?: boolean
  data: Record<string, any>
}

/**
 * 虚拟视频角色
 */
export interface VreoVideo {
  duration: number
  start: number
  end: number
  url: string
  avatar?: {
    url?: string
    force?: true
  }
}

/**
 * 剧本结构
 */
export interface VreoUnit {
  categoryId?: string
  categoryText?: string
  frontRequestId?: string
  index?: string | number
  configure?: { pageTitle?: string }
  keyframes: VreoKeyframe[]
  video: VreoVideo
}



/**
 * 剧本事件
 */
export type VreoKeyframeEvent = { [key in VreoKeyframeEnum]: (keyframe: VreoKeyframe, currentTime: number) => void } & {
  // 新数据载入完成
  loaded: (vreoUnit: VreoUnit) => void
  // 语音状态 播放 -> 暂停
  paused: (ended?: boolean) => void
  // 语音状态 暂停 -> 播放
  playing: () => void
  // 未知的 剧本帧类型
  unknownKeyframeType: (keyframe: Record<string, any>) => void
  // 一次播放完成（试用，只抛事件不做任何额外的逻辑处理）
  ended: () => void
}